requirements.txt
altgraph==0.17.5
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
distro==1.9.0
Flask==3.1.2
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.7.1
greenlet==3.2.4
idna==3.11
invoke==2.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
paramiko==4.0.0
posthog==7.4.2
psutil==7.1.3
psycopg2-binary==2.9.11
pycparser==2.23
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
PyJWT==2.10.1
PyMySQL==1.1.2
PyNaCl==1.6.1
pyodbc==5.3.0
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
setuptools==80.9.0
shiboken6==6.10.1
six==1.17.0
SQLAlchemy==2.0.44
sshtunnel==0.4.0
typing_extensions==4.15.0
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3